requirements.txt
anyio==3.6.1
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
cycler==0.11.0
fastapi==0.79.0
FastAPI-SQLAlchemy==0.2.1
fonttools==4.34.4
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
joblib==1.1.0
kiwisolver==1.4.4
matplotlib==3.5.2
numpy==1.23.1
packaging==21.3
passlib==1.7.4
Pillow==9.2.0
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
threadpoolctl==3.1.0
typing_extensions==4.3.0
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
